pyproject.toml
[project]
name = "mm-bmad-v2-backend"
version = "0.1.0"
description = "Backend API for showcase gallery and booking portal"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    # Dependencies will be added via `uv add` command
    # FastAPI, SQLAlchemy, etc. will be added in Epic 2
    "bcrypt>=5.0.0",
    "cloudinary>=1.44.1",
    "fastapi>=0.128.0",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.45",
    "uvicorn[standard]>=0.40.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

